refactor(client): Move all entity client tests to same file
